requirements.txt
pinecone>=3.0.0
sentence-transformers>=2.2.2
rank-bm25>=0.2.2
python-dotenv>=1.0.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
scikit-learn>=1.3.0
numpy>=1.24.0
textblob>=0.17.0
symspellpy>=0.0.1